Switch to using Oracle's packaged plist file
This is a fix for issue #27 opened by navels. Oracle now includes a
working plist file in /Library/LaunchDaemons, so we have removed the
addition of our plist file during the setup script, as it was redundant.
MySQL is also set to run on startup by default, which can be changed
from the MySQL prefpane.